+++ src/json-util.cpp
@@ -418,6 +418,7 @@ void ju_get_opt_field(const nlohmann::json &j, const K &key, machine_runtime_con
     }
     ju_get_field(j[key], "concurrency"s, value.concurrency, path + to_string(key) + "/");
     ju_get_opt_field(j[key], "skip_root_hash_check"s, value.skip_root_hash_check, path + to_string(key) + "/");
+    ju_get_opt_field(j[key], "skip_version_check"s, value.skip_version_check, path + to_string(key) + "/");
 }
 
 template void ju_get_opt_field<uint64_t>(const nlohmann::json &j, const uint64_t &key, machine_runtime_config &value,
@@ -1196,6 +1197,7 @@ void to_json(nlohmann::json &j, const machine_runtime_config &runtime) {
     j = nlohmann::json{
         {"concurrency", runtime.concurrency},
         {"skip_root_hash_check", runtime.skip_root_hash_check},
+        {"skip_version_check", runtime.skip_version_check},
     };
 }